1958 corrections
1) 26 March 1958: NBC should be moved down, from header to description, as in my note.

2) 3 July 1958.
You have only partially applied my corrections, I can't say if it you did it on purpose.
The description preceding the first three selections should be: Afternoon radio broadcast. Newport Jazz Festival, Freebody Park.
Take The A Train should be moved up as the first title.
The description preceding the remaining selections could be Evening concert. Newport Jazz Festival, Freebody Park.
Otherwise, with the same text on both parts of the session, why have you split it?

3) 7 November 1958: Add Sv 1018540 to these Selections
58-11-07-002 Take The "A" Train DE5850a
58-11-07-003 Newport Up DE5850b
58-11-07-004 My Funny Valentine DE5850c
58-11-07-005 Perdido DE5850d
58-11-07-006 Sophisticated Lady DE5850e
58-11-07-007 Sonnet To Hank Cinq DE5850f
58-11-07-008 What Else Can You Do With A Drum? - vOB DE5850g

4) 14 November (DE5853).
Instead of adding to the Medley header, the CD boxset Sv 108-8617 should be added to each component of the medley, just like Sv STCD-8324.

This is 1958.