Correct take number for DE2803e.
From SL

You show Take It Easy as take 1.

The actual take number for this master is unknown.

ESLTO and Jubilee Stomp are both take one. We know this by reference to English
Pathe Actuelle 11558, a flush label pressing on which the master numbers and
takes are visible "under the label" as inscribed by the engineer at the
recording session. Unfortunately, Take It Easy was only issued on U.S. Pathe and
Perfect, which are "sunken" label pressings that don't bear matrix number or
take data.