Fixes for 21 Feb and 20 March 1930
based on today's mails from David Palmquist and Steven Lasker, the following corrections should be inserted in ellingtonia.com

1) Session DE3002, 21 February 1930.

Add the venue to the description:
Brunswick recording session. Brunswick studio, 799 Seventh Ave.

In the band personnel, add "probably" or "possibly" before Irving Mills(v),
based on Steven Lasker's comment: "since we can't hear When You're Smiling mx. E32212-A-B (no metal parts or tests are known to survive), we can't say who sang the vocal on this version; the surviving files don't tell us."

Replace the titles Maori with Maori (A Samoan Dance) on both lines 30-02-21-002 and 30-02-21-003.

Correct the Matrix Nr E32210-A as E32210-A or -B on the line 30-02-21-002:
30-02-21-002     E32210-A or -B       Maori (A Samoan Dance)       Br 4776, Cl(F) 586, COPS 7395-D1, De GRD-3-640...           DE3002b

2) Session DE3004, 20 March 1930.

Add the venue to the description:
Brunswick recording session. Brunswick studio, 799 Seventh Ave.

Replace the titles Maori with Maori (A Samoan Dance) on both lines 30-03-20-003 and 30-03-20-004.

Correct the Matrix Nr E32449-A as E32449-A or -B on the line 30-03-20-005:
30-03-20-005     E32449-A or -B       Admiration                          Br 4776, Cl(F) 586, COPS 7395-D1, De GRD-3-640...            DE3004e