Remove vocal from DE6374n
From Brian K

DE6374n (Take The "A" Train from 12 October 1963) states an Ernie Shepherd
vocal. There is none. Also, it uses the theme arrangement of Take The "A" Train.
The audio file for the session was recently posted in the Goodies section at
Ellington Galaxy.

Additionally, there appear to be a few bars from "Frolic Sam" played by
Ellington and Hodges (unlikely as it might seem) just prior to "Tutti For
Cootie"