From EC
In the 29 Dec 26 session, a slight correction: "Unknown as/cl: as per
note on 29 November 1926" would be better Unknown alto sax as per
note...", because it is n/k(as), not (as,cl). ("Unknown as as" would
sound ugly.)

In all the sessions from 9 January 1928 to 10 July 1928 please remove ss
from Carney's instruments: replace Harry Carney(cl,ss,as,bar); with
Harry Carney(cl,as,bar).

In the session 20 Dec 1928, replace Greer(d) with Greer(d,v) in the band
members list.

Regarding Hardwick instruments, we can leverage from the booklets of RCA
Centennial Edition and Complete Brunswick and Vocalion 1926-31 (both by
Steven Lasker, who else?),  for the following sessions:

06 October 1927: Otto Hardwick(as,bar) RCA
26 October 1927: Otto Hardwick(cl,as) RCA
19 December 1927: Otto Hardwick(as,bar) RCA
29 December 1927: Otto Hardwick(as,bsx) Br/Vo
21 March 1928: Otto Hardwick(as) Br/Vo
26 March 1928: Otto Hardwick(as) RCA

Less reliable sourced for the following sessions:
03 November 1927; 09 January 1928; 19 January 1928; c. 07 March 1928.
Let's keep the NDESOR long list of Hardwick instruments.