1932 correction
****** Correction. ******

1932 09 21. Jive Stomp should instead read Jive.

Steven Lasker's explanation:
RCA's file sheet for this session lists the second title as "Jive Fox Trot Comp--Duke Ellington." Neither metal parts nor pressings are known for this master, which is presumed lost.
Ellington composed, and on 1933 08 15 recorded, a piece called Jive Stomp. The two recordings are not necessarily of the same piece.