1933 corrections
09 May: replaced release Br 6607 with Br(E) 01527 with flag "◊".
Steve Lasker: «Get Yourself a New Broom was released on Br 6607 on 1933 07 15. The first issue of this master was on Br(E)01527A in early June 1933. I've never found release dates for British Brunswick, but the printed program for Ellington's Melody Maker concert at the Palladium (on 1933 06 12) contains an ad for Br(E)1527, so this issue was evidently the first.»

04 March:  We don't have an exact date for this film, so the date has been preceded by "Circa".