1967
"The Biggest" (short form of "The Biggest and Busiest Intersection") has been replaced with Kixx in all its occurrences. Kixx is the title of the tune as copyrighted in 1967. The same tune appeared with other titles, "Traffic Jam", "WARR". Finally, when it was included in the Second Sacred Concert, it was given the title "The Biggest and Busiest Intersection", by which it is best known.

Here, all occurrences of "The Biggest" are replaced with Kixx.

"The Biggest (Traffic Jam)" becomes "Kixx [Traffic Jam]"
"The Biggest (Kixx)" becomes "Kid"
"The Biggest (Warr)" becomes "Kixx [Warr]"

-----

All alternative titles follow main title in square brackets [] instead of parentheses ().