1969
"The Biggest" has been replaced with "The Biggest and Busiest Intersection [Kixx]" in all its occurrences. The tune was copyrighted as Kixx in 1967 and was later given the title "The Biggest and Busiest Intersection" when it was included in the Second Sacred Concert.

27 September 1969 (concerts DE6937 and DE6938): replaced n7k choir with The Westminster Choir.

-----

All alternative titles follow main title in square brackets [] instead of parentheses ().