1933
EC:
Here is 1933. With respect to the current version, attached, its only change is Joe Garland(ts,bar) on 9 May, instead of Joe Garland(ts).

My doubt was: which alto player was out on the first song, Happy as the Day is Long, of the 9 May session? Carney or Hardwick?

Steven wrote:
The ledger shows 3 saxes on Happy as the Day is Long, and 4 on the other titles. Since the tenor sax solo heard on Happy is consistent with Garland's recordings with the Mills Blue Rhythm Band, and Hodges solos on this title, and section alto is played by Carney or Hardwick, but so little it's not possible to say which man was present.
JH, ss (on Raisin' the Rent)/as; OH, as (out on Happy as the Day?); HC, as/bar; Joe Garland, ts/bar. Had Carney played on Happy, you'd expect he'd be given the baritone solo on that title instead of Garland, which suggests that Hardwick is likely the section altoist on this title.

So, Carney is probably out on Happy as the Day as already written in ellingtonia.com